Question:01
Create a simple Car class that includes attributes like name, model, and year.
Create two derived classes, ElectricCar and GasCar, which inherit from the Car class.
Consider adding a battery capacity attribute to the ElectricCar class and a fuel efficiency
characteristic to the GasCar class.The user will enter the battery capacity and fuel efficiency
attribute values based on the type of vehicle. For electric cars, the input should be battery
capacity, while for gas cars, the input should be fuel efficiency.